feat: 공통 에러 다이알로그 함수 생성, 에러 타입 파일 추가, 에러 핸들링 용 atom 생성
