--- src/stores/common-atom.ts
@@ -1,10 +1,0 @@
-import { atom } from "jotai";
-
-export const isErrorDialogOpenAtom = atom<boolean>(false);
-export const errorDialogAtom = atom<{
-  isOpen: boolean;
-  heading: string;
-  body: string;
-  button: string;
-  redirectUrl?: string;
-} | null>(null);
