+++ src/stores/error-atom.ts
@@ -1,0 +1,5 @@
+import type { ErrorDialogType } from "@/@types/error";
+import { atom } from "jotai";
+
+export const isErrorDialogOpenAtom = atom<boolean>(false);
+export const errorDialogAtom = atom<ErrorDialogType | null>(null);
